+++ Hinotori/Administrador/src/main/java/Administrador/model/dao/ClienteDao.java
@@ -8,8 +8,8 @@ public interface ClienteDao {
 	
 	void insert(Cliente obj);
 	void update(Cliente obj);
-	void delete(Integer id);
-	Cliente find(Integer id);
+	void delete(Long id);
+	Cliente find(Long id);
 	List<Cliente> findAll();
 
 }
