+++ Hinotori/Administrador/src/main/java/Administrador/model/dao/EstadoDao.java
@@ -8,8 +8,8 @@ public interface EstadoDao {
 
 	void insert(Estado obj);
 	void update(Estado obj);
-	void delete(Integer id);
-	Estado find(Integer id);
+	void delete(Long id);
+	Estado find(Long id);
 	List<Estado> findAll();
 	
 }
